Change App and main to jsx
