+++ client/index.html
@@ -17,7 +17,7 @@
     <div id="root"></div>
     <script
       type="module"
-      src="/src/main.tsx"
+      src="/src/main.jsx"
     ></script>
   </body>
 </html>
