--- client/src/main.tsx
@@ -1,20 +1,0 @@
-import { StrictMode } from "react";
-import { createRoot } from "react-dom/client";
-
-import { ChakraProvider, extendTheme } from "@chakra-ui/react";
-
-import App from "./App.tsx";
-
-const colors = {
-  brand: {},
-};
-
-const theme = extendTheme({ colors });
-
-createRoot(document.getElementById("root")!).render(
-  <StrictMode>
-    <ChakraProvider theme={theme}>
-      <App />
-    </ChakraProvider>
-  </StrictMode>
-);
